{"types":"C","runs":95}

+++ Assignment2Pyramid.java
@@ -12,9 +12,9 @@ public class Assignment2Pyramid extends GraphicsProgram {
 	}
 
 	private void addBase() {
-	for (int i = 14; i > 1; i--) {
-		for (int j = 1; j<=BRICKS_IN_BASE; j++) {   	
-				double x = getWidth()/2 - (BRICK_WIDTH * i/2) +(i * BRICK_WIDTH )  ;
+	for (double i = 14; i > 1; i--) {
+		for (double j = 1; j<=BRICKS_IN_BASE; j++) {   	
+				double x = getWidth()/2 - (BRICK_WIDTH * i/2) +(i * BRICK_WIDTH )   ;
 				double y = getHeight() - j* BRICK_HEIGHT;
 				GRect brick = new GRect(x, y, BRICK_WIDTH, BRICK_HEIGHT);
 				add(brick);
